requirements.txt
altgraph==0.17.2
future==0.18.2
numpy==1.22.2
pandas==1.4.1
pefile==2021.9.3
Pillow==9.0.1
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.1
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
six==1.16.0
